backend/requirements.txt
asgiref==3.7.2
Django==4.0.3
sqlparse==0.4.4
typing_extensions==4.9.0
mysqlclient
pymysql
cryptography
djangorestframework
jwt
PyJWT